Move autocomplete util to main module
